backend/requirements.txt
flask==3.0.3
pyjwt==2.10.1
cryptography==44.0.1
